coresight_requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-generator==1.10
attrs==24.2.0
backoff==2.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.20.0
exceptiongroup==1.2.2
geojson==3.1.0
geojson-pydantic==1.2.0
geopandas==1.0.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
loguru==0.7.2
morecantile==5.4.2
#numpy==2.1.3
packaging==24.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyogrio==0.10.0
pyparsing==3.2.0
#pyproj==3.7.0
#pystac==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rasterio==1.4.1
requests==2.32.3
rio-cogeo==5.3.6
rio-stac==0.10.1
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3

